requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
faker==20.1.0
